First simulator draft
Extremely rough draft, but it does a thing or two. Namely it can
initialize a simulator and move back and forth. Currently all output
happens in the text channel and there is no way to choose a transition.
But it will be mostly a matter of replicating the React app from recipe