Fix type inference of local variables
I did it! As it turns out, if you peek the .ref field of a Reference<>
object that does not link, Langium gets really mad.

We use an inference rule to make type inference of these nodes "lazy",
and lookup if a custom type (ie., an Enum) exists that matches the
ref's .$refText field (which is always available).